move file store
